requirements.txt
fastapi==0.71.0
uvicorn==0.16.0
tortoise-orm==0.19.0
orjson==3.6.5
loguru==0.5.3
aiomysql==0.0.22
async-lru==1.0.2
pyjwt==2.3.0
requests==2.28.2